requirements.txt
# Lambda WSGI adapter
serverless-wsgi>=3.0.0

# Web app
flask>=3.0.0
python-dotenv>=1.0.0

# AWS
boto3>=1.28.0

# LangChain & RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-aws>=0.0.20
langchain-core>=0.1.0

# Vector store
chromadb>=0.4.0

# PDF (for upload / ingestion)
pypdf>=3.0.0

# Utilities
tqdm>=4.65.0
